added unit test for users store module and jest matchers
This checkin uses the leasons learned on unit testing the auth store, to test user store. This showed me exactly why Typescript is necessary. I had to redefine the user interface that the auth store defined because now the user store uses the same type but added fields. I had no idea those fields should be part of the user interface until I saw it used in the user module. I shouldn't have to hunt all over the source code to find out what the shape of a type is supposed to look like. Typescript gives you that advantage when you are coming in and trying to learn a code base.

I also added type definitions for the jest matchers so that I could use them from typescript unit tests.