converted the unit tests for the components to TS
I didnt touch the components because I figure the value isn't in the components. Any person building an enterprise apps will delete the components and use something like Vuetify. However, I did cleanup and convert the component unit tests because the unit tests can be an illustration on how to test components. However, even that is of limited value because if you are using typescript, you should be using the composition API and most of your unit testing will be focused on the modules you import into your component as their view models.