tweaks and docs update
This change was to start to enhance the documentation. The readme changed to explain the purpose of this project and how it relates to Chris' original project.

testing was augmented with information I wish I didn't have to figure out myself.

and some settings have been tweaked:
eslintrc: I added some comments and I am trying to track down (unsuccessfully) why it wont reformat on save yet.
packages.json: I updated ts-jest with the version that matches the version of jest we are using. I have not upgraded to the latest version of jest yet because there seems to be a performance problem with the corresponding ts-jest.
I have updated the sfc snippet, but i have to figure out to invoke it.
I have updated the component generator but I want to check in my changes before I test it because I need to revert the repo after testing it.

the code changes are just trying to fix up small errors that still exist.