converted component generator
the component generator now creates a composition API based component and the unit test
is now a typescript file. I also fixed up a few things based on what I
tend to always add to components. For instance, the props field, even
when empty needs to be there for typescript to recognize that it is a
proper module for vue-test-utils. So I added that (with a comment
indicating why an empty props needs to be retained.
I also set the name

I changed the template from a <div \> to a proper open and close
because I feel that most components will start with a div, or if they
aren't going to use a div, when they change the div tag to something
else, the end tag will change too. It just seems to provide more
assistance to the user.