updated module generator
I changed the module generator to output typescript files and to provide
more assistance in setting up a module. For instance, it is quite likely that
a unit test will always have to create an instance of a store, so I added
creating an instance of the store to the unit test.
Also, I believe that adding Arrange, Act and Assert comments add to
the readability of unit tests, so the generator adds those comments
so it becomes easier for users to follow that structure.

For the module, I added more help and guidance as well. Each module
should export an interface that describes the shape of the state data.
Then the state should be typed to that interface. Also it because easier
to write unit tests when you export a store that is all the items together
so that is automatically added.